requirements.txt
anime-pgen==1.0.1
annotated-types==0.6.0
anyio==4.3.0
boto3==1.34.79
botocore==1.34.79
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.6
colour==0.1.5
fastapi==0.110.1
h11==0.14.0
httptools==0.6.1
idna==3.6
jmespath==1.0.1
joblib==1.3.2
numpy==1.26.4
Pillow==9.5.0
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
ratelimiter==1.2.0.post0
reactivex==4.0.4
requests==2.31.0
s3transfer==0.10.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typer==0.4.2
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
